--- src/main/java/com/project/doubleshop/domain/member/entity/Authority.java
@@ -1,13 +1,0 @@
-package com.project.doubleshop.domain.member.entity;
-
-/**
- * CONSUMER: 소비자
- * SELLER: 판매자
- * ADMIN: 관리자
- */
-@Deprecated
-public enum Authority {
-    CONSUMER,
-    SELLER,
-    ADMIN;
-}
