--- src/main/java/com/project/doubleshop/domain/member/entity/Grade.java
@@ -1,13 +1,0 @@
-package com.project.doubleshop.domain.member.entity;
-
-/**
- * NORMAL: 일반 회원
- * SILVER: 실버 회원
- * GOLD: 골드 회원
- */
-@Deprecated
-public enum Grade {
-    NORMAL,
-    SILVER,
-    GOLD;
-}
