--- src/main/java/com/project/doubleshop/domain/member/entity/Status.java
@@ -1,13 +1,0 @@
-package com.project.doubleshop.domain.member.entity;
-
-/**
- * NORMAL: 일반 사용자 계정
- * DORMANT: 휴면 계정
- * BANNED: 사용 정지된 계정
- */
-@Deprecated
-public enum Status {
-    NORMAL,
-    DORMANT,
-    BANNED;
-}
