--- src/main/java/com/project/doubleshop/domain/member/entity/Type.java
@@ -1,13 +1,0 @@
-package com.project.doubleshop.domain.member.entity;
-
-/**
- * REGULAR: 일반 사용자
- * NAVER: 네이버 계정 사용자
- * KAKAO: 카카오 계정 사용자
- */
-@Deprecated
-public enum Type {
-    REGULAR,
-    NAVER,
-    KAKAO;
-}
