--- src/main/java/com/project/doubleshop/web/config/LogInMemberArgumentResolver.java
@@ -1,33 +1,0 @@
-package com.project.doubleshop.web.config;
-
-import com.project.doubleshop.domain.member.service.LogInService;
-import com.project.doubleshop.web.member.annotation.CurrentMember;
-import lombok.RequiredArgsConstructor;
-import org.springframework.core.MethodParameter;
-import org.springframework.stereotype.Component;
-import org.springframework.web.bind.support.WebDataBinderFactory;
-import org.springframework.web.context.request.NativeWebRequest;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.method.support.ModelAndViewContainer;
-
-@Component
-@RequiredArgsConstructor
-public class LogInMemberArgumentResolver implements HandlerMethodArgumentResolver {
-
-    private final LogInService logInService;
-
-    @Override
-    public boolean supportsParameter(MethodParameter methodParameter) {
-        return methodParameter.hasParameterAnnotation(CurrentMember.class);
-    }
-
-    @Override
-    public Object resolveArgument(MethodParameter methodParameter,
-                                  ModelAndViewContainer modelAndViewContainer,
-                                  NativeWebRequest nativeWebRequest,
-                                  WebDataBinderFactory webDataBinderFactory)
-            throws Exception {
-        return logInService.getLogInMember();
-    }
-
-}
