--- src/main/java/com/project/doubleshop/web/config/WebConfig.java
@@ -1,33 +1,0 @@
-package com.project.doubleshop.web.config;
-
-import com.project.doubleshop.web.member.interceptor.LogInCheckInterceptor;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import java.util.List;
-
-/**
- * WebMvcConfigurer: 기본 MVC 설정에 추가적인 기능을 커스터마이징 할 수 있다.
- */
-@Configuration
-@RequiredArgsConstructor
-public class WebConfig implements WebMvcConfigurer {
-
-    private final LogInCheckInterceptor logInCheckInterceptor;
-
-    private final LogInMemberArgumentResolver logInMemberArgumentResolver;
-
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(logInCheckInterceptor);
-    }
-
-    @Override
-    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(logInMemberArgumentResolver);
-    }
-
-}
