--- src/main/resources/mapper/memberMapper.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper
-        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-<mapper namespace="com.project.doubleshop.domain.member.mapper.MemberInfoMapper">
-
-    <insert id="insertMember" parameterType="com.project.doubleshop.web.member.dto.MemberSaveRequestDto">
-        INSERT INTO member (user_id, password, name, phone, email, status)
-        VALUES (#{userId}, #{password}, #{name}, #{phone}, #{email}, #{status})
-    </insert>
-
-    <select id="selectUserId" resultType="com.project.doubleshop.web.member.dto.MemberInfoDto">
-        SELECT user_id
-        FROM member
-        WHERE user_id = #{userId}
-        LIMIT 1
-    </select>
-
-    <select id="selectEmail" resultType="com.project.doubleshop.web.member.dto.MemberInfoDto">
-        SELECT email
-        FROM member
-        WHERE email = #{email}
-        LIMIT 1
-    </select>
-
-    <select id="selectByUserId" resultType="com.project.doubleshop.domain.member.entity.Member">
-        SELECT user_id, name, phone, email
-        FROM member
-        WHERE user_id = #{userId}
-    </select>
-
-    <select id="selectByUserIdAndPassword" resultType="com.project.doubleshop.web.member.dto.LogInRequestDto">
-        SELECT user_id, password
-        FROM member
-        WHERE user_id = #{userId}
-        AND password = #{password}
-    </select>
-
-</mapper>
